initial separation of content settings and domain settings
