+++ domain-server/resources/web/base-settings-scripts.html
@@ -1,0 +1,7 @@
+<script src='/js/underscore-min.js'></script>
+<script src='/js/underscore-keypath.min.js'></script>
+<script src='/js/bootbox.min.js'></script>
+<script src='/js/form2js.min.js'></script>
+<script src='/js/bootstrap-switch.min.js'></script>
+<script src='/js/shared.js'></script>
+<script src='/js/base-settings.js'></script>
