+++ domain-server/resources/web/content/index.shtml
@@ -1,45 +1,19 @@
 <!--#include virtual="header.html"-->
 
-<div class="col-md-10 col-md-offset-1">
-  <div class="row">
-    <div class="col-xs-12">
-      <div class="alert" style="display:none;"></div>
-    </div>
-  </div>
+<script type="text/javascript">
+   var Settings = {
+      content_settings: true,
+      endpoint: "/content-settings.json",
+      path: "/content/"
+   };
+</script>
 
-  <div class="row">
-    <div class="col-xs-12">
-      <div class="panel panel-default">
-        <div class="panel-heading">
-          <h3 class="panel-title">Upload Entities File</h3>
-        </div>
-        <form id="upload-form" action="upload" enctype="multipart/form-data" method="post">
-          <div class="panel-body">
-            <p>
-              Upload an entities file (e.g.: models.json.gz) to replace the content of this domain.<br>
-              Note: <strong>Your domain's content will be replaced by the content you upload</strong>, but the backup files of your domain's content will not immediately be changed.
-            </p>
-            <p>If your domain has any content that you would like to re-use at a later date, save a manual backup of your models.json.gz file, which is usually stored at the following paths:</p>
-            <label class="control-label">Windows</label>
-            <pre>C:/Users/[username]/AppData/Roaming/High Fidelity/assignment-client/entities/models.json.gz</pre>
-            <label class="control-label">OSX</label>
-            <pre>/Users/[username]/Library/Application Support/High Fidelity/assignment-client/entities/models.json.gz</pre>
-            <label class="control-label">Linux</label>
-            <pre>/home/[username]/.local/share/High Fidelity/assignment-client/entities/models.json.gz</pre>
-            <br>
-            <input type="file" name="entities-file" class="form-control-file" accept=".json, .gz">
-            <br>
-          </div>
-          <div class="panel-footer">
-            <input type="submit" class="btn btn-info" value="Upload">
-          </div>
-        </form>
-      </div>
-    </div>
-  </div>
-</div>
+<!--#include virtual="base-settings.html"-->
 
 <!--#include virtual="footer.html"-->
-<script src='js/content.js'></script>
-<script src='/js/sweetalert.min.js'></script>
+
+<!--#include virtual="base-settings-scripts.html"-->
+
+<script src="js/content.js"></script>
+
 <!--#include virtual="page-end.html"-->
